Generator_Synth/music_gen/requirements.txt
numpy
openai
torch
torchaudio
midiutil
librosa
tensorflow
tensorflow_hub
csv
pandas
IPython
scipy
diffusers
pillow
pydub
